cron/upload_yt_shorts/requirements.txt
google-api-python-client~=2.24.0
google_auth_oauthlib~=0.4.6
pytube~=11.0.1
moviepy~=1.0.3
requests~=2.26.0
python-dotenv~=0.19.0